requirements.txt
bidict==0.21.2
blinker==1.4
cffi==1.14.6
click==8.0.1
cobble==0.1.3
colorama==0.4.4
corha==0.3.1
dnspython==2.1.0
dotmap==1.3.24
email-validator==1.1.3
Flask==2.0.1
Flask-Login==0.5.0
Flask-Navigation==0.2.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
gevent==21.8.0
gevent-websocket==0.10.1
greenlet==1.1.1
h11==0.12.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
mammoth==1.4.16
Markdown==3.3.4
MarkupSafe==2.0.1
Pillow==8.3.1
psycopg2-binary==2.9.1
pycparser==2.20
python-engineio==4.2.1
python-socketio==5.4.0
SQLAlchemy==1.4.22
Werkzeug==2.0.1
wsproto==1.0.0
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.4.0
pyScss